refactor(mcp): split tool registration into per-module register()
Moves the 12 nested @mcp.tool() wrappers out of build_server into a
register(mcp, context, *, transport) function in each tool module.
build_server is now a short list of register() calls with no C901 noqa.
Adds transport param (unused, ARG001-suppressed) ahead of Task 2.
